Add files via upload
Risk equals vulnerabilities times threats, times consequences, would be a way of representing it in a formula. The likelihood of threats exploiting vulnerabilities in your security controls and the consequences of that is known as risk. Risk to your assets, the risk to you, the risk to your privacy and anonymity. And the threats and adversaries that you face, these are called your threat landscape, or your threat model. Your threat landscape will be individual to you.